Fix join elimination confusion between columns of same name
Eliminating a referential integrity join requires the foreign key to
reference the unique key somehow.

The original solution was to use an UnresolvedAttribute and modify the
analyzer to do the resolution. However, this required rewriting the
referencing attribute in the foreign key whenever the referenced
attribute in the unique key was rewritten, such as to resolve a
self-join. Modifying Spark was not a good solution.

The next solution was to leave the referencing attribute unresolved
until optimization time, then attempt to resolve it against each
candidate parent table. However, this could produce false positives when
different columns were named the same thing (see added unit test in
KeyHintSuite).

This commit adds a stable identifier to each UniqueKey which is used as
the reference. This way, the identifier is not affected by attribute
rewrites.